Improve flag descriptions
When dumping, we want to be able to specify
the kind of record to dump. A record type roughly
corresponds to one table. The detail is that
a record is not simply the rows from a table,
but also the rows from other tables that the table
has foreign key relationships with.

--number looks like a better description for the flag to specify number of rows.

A sample command would like this:

klepto \
  --from 'root:root@tcp(localhost:3306)/fromDB' \
  --to 'root:root@tcp(localhost:3306)/toDB' \
  --primary-record-type 'user' \
  --number '10'