Refactor GetPreamble() to separate out hostname retrieval
We want to write a function to return all relationships
stored on a table. We will later use this function to enable us
determine dependencies between record types. In the new function
we want to write, we will retrieve the database name. There is
already a way to do this in GetPreamble(). So move that implementation
out and while doing that, do something similar for retrieving hostname.