Configure record type
We want to use only the primary record for generating
the query if the user specifies that in the configuration.
Otherwise, we want to get tables from the db and dump everything.
This commit implements that logic integrates the
function that reads the primary_record_type configuration
from the config file. If a number is provided, we
limit the query that fetches records to that number.
We can configure record type like this

[primary_record_type]
"users" = 1000