Wire reading tables in
We want to use only the primary record for generating
the query if the user specifies that in the configuration.
Otherwise, we want to get tables from the db and dump everything.
This commit implements that logic integrates the
function that reads the primary_record_type configuration
from the config file.
We only dump the structure in this commit.
In the next commit, we hope to also create the
INSERT statements for only the specified tables.