Add --data-only to copy data without structure
This was prompted as a workaround for the issue presented in #139.
By allowing data-only transfers to be run, we can use multiple config
files in order to extract multiple subsets from the same table. An
initial run, without `--data-only`, copies the DB structure and
anonmized users while a second run using `--data-only` and a different
config file copies unmodified versions of admin users to the same table.

It's likely useful in many other caeses as well.